Hi-res: widen visplane top/bottom to u16 (fix floor/ceiling/sky); EE serial log
Visplane fix: visplane_t.top[]/bottom[] were byte arrays (0-255) with 0xff
= 'no span'. At 640x400 the per-column clip-Y exceeds 255 -> wraps, and
Y=255 collides with the sentinel -> corrupt floor/ceiling/sky spans, the
'R_MapPlane: .. at 255' crash, and the blurry/missing flats. Widened to
unsigned short (sentinel 0xffff); r_plane.c 0xff -> 0xffff. (byte memsets
stay -- 0xff byte-fill yields 0xffff per short.)

EE serial console: _write (ps2_bootscr.c) used to drop EE stdout once the
boot screen closed. Now it always forwards fd1/2 to SIO (sio_write), which
PCSX2 mirrors to its log -- so engine output is visible remotely. ps2_pad.c
logs button edges + analog direction changes (PAD_LOG) through it.

+++ doomgeneric/r_defs.h
@@ -431,16 +431,17 @@ typedef struct
   int			maxx;
   
   // leave pads for [minx-1]/[maxx+1]
-  
-  byte		pad1;
-  // Here lies the rub for all
-  //  dynamic resize/change of resolution.
-  byte		top[SCREENWIDTH];
-  byte		pad2;
-  byte		pad3;
-  // See above.
-  byte		bottom[SCREENWIDTH];
-  byte		pad4;
+
+  // top[]/bottom[] hold per-column screen-Y clip values with 0xffff = "no span".
+  // Widened from byte to unsigned short: a byte only reaches 255, but a hi-res
+  // (640x400) view has Y up to 399, so the values wrapped AND collided with the
+  // old 0xff sentinel -> corrupt floor/ceiling/sky spans + R_MapPlane crash.
+  unsigned short	pad1;
+  unsigned short	top[SCREENWIDTH];
+  unsigned short	pad2;
+  unsigned short	pad3;
+  unsigned short	bottom[SCREENWIDTH];
+  unsigned short	pad4;
 
 } visplane_t;
